Make protocol.h self-contained, and compile it in CI
The firmware build got past dependency resolution and into the compiler,
where it failed on the first translation unit that includes protocol.h
without Arduino.h ahead of it:

    src/core/protocol.h:103: error: 'size_t' has not been declared

The header used size_t while including only <stdint.h>. Every other file
happened to include Arduino.h first and got it transitively; vad.cpp did
not. Added <stddef.h> - a header has to stand on its own.

Rather than fix that blind and wait another CI cycle to find the next one,
protocol.h is now compiled and executed as part of the test suite.
firmware/test/test_protocol_host.cpp includes it first, with nothing before
it, and is built with g++ -Wall -Wextra -Werror. It is plain C++ with no
Arduino dependency, so no Xtensa toolchain is needed.

That immediately paid for itself by catching a second defect in the same
header: walleParseBinHeader compared an int against a size_t. Harmless at
these magnitudes, wrong in principle, and now cast properly.

The harness also runs the real C encoder and decoder against the Python
ones. Comparing #define values catches drift in the constants but not in
the code around them, and a mismatch there is garbled audio on hardware.
Eight new tests confirm the two produce byte-identical frames across empty,
short, full-audio-frame and 4 KB JPEG payloads, that C parses what Python
emits, and that both reject bad magic and truncation identically.

Also host-compiled vad.cpp, the file that first failed, to confirm the fix
directly. 156 tests pass; ruff clean.

+++ firmware/src/core/protocol.h
@@ -30,6 +30,10 @@
 // ---------------------------------------------------------------------------
 #pragma once
 
+// Both are needed and neither is optional: size_t comes from <stddef.h>, and a
+// translation unit that includes this header *before* Arduino.h (as vad.cpp
+// does) gets nothing transitively. A header has to stand on its own.
+#include <stddef.h>
 #include <stdint.h>
 
 // Bump only for breaking changes; the server refuses mismatched majors.
@@ -108,5 +112,8 @@ inline bool walleParseBinHeader(const uint8_t *in, size_t inLen,
   *flags = in[2];
   *seq = static_cast<uint16_t>(in[4] | (in[5] << 8));
   *len = static_cast<uint16_t>(in[6] | (in[7] << 8));
-  return (*len + WALLE_BIN_HEADER_LEN) <= inLen;
+  // Cast before adding: *len promotes to int, and comparing that against a
+  // size_t is a signed/unsigned comparison the compiler is right to warn
+  // about. Harmless at these magnitudes, wrong in principle.
+  return (static_cast<size_t>(*len) + WALLE_BIN_HEADER_LEN) <= inLen;
 }
